+++ ERPCore2/Services/Sales/SalesDeliveryService.cs
@@ -351,14 +351,14 @@ namespace ERPCore2.Services
                         : allTransactions.Where(t => !t.TransactionNumber.EndsWith("_DEL")).ToList();
 
                     // 建立已處理過庫存的明細字典（ProductId + WarehouseId + LocationId -> 已處理庫存淨值）
-                    var processedInventory = new Dictionary<string, (int ProductId, int WarehouseId, int? LocationId, int NetProcessedQuantity)>();
+                    var processedInventory = new Dictionary<string, (int ProductId, int WarehouseId, int? LocationId, decimal NetProcessedQuantity)>();
                     
                     foreach (var trans in existingTransactions)
                     {
                         var key = $"{trans.ProductId}_{trans.WarehouseId}_{trans.WarehouseLocationId?.ToString() ?? "null"}";
                         if (!processedInventory.ContainsKey(key))
                         {
-                            processedInventory[key] = (trans.ProductId, trans.WarehouseId, trans.WarehouseLocationId, 0);
+                            processedInventory[key] = (trans.ProductId, trans.WarehouseId, trans.WarehouseLocationId, 0m);
                         }
                         // 累加所有交易的淨值（注意：出庫的 Quantity 是負數）
                         var oldQty = processedInventory[key].NetProcessedQuantity;
@@ -368,7 +368,7 @@ namespace ERPCore2.Services
                     }
                     
                     // 建立當前明細字典
-                    var currentInventory = new Dictionary<string, (int ProductId, int? WarehouseId, int? LocationId, int CurrentQuantity)>();
+                    var currentInventory = new Dictionary<string, (int ProductId, int? WarehouseId, int? LocationId, decimal CurrentQuantity)>();
                     
                     foreach (var detail in currentDelivery.DeliveryDetails)
                     {
@@ -392,13 +392,13 @@ namespace ERPCore2.Services
                         var hasCurrent = currentInventory.ContainsKey(key);
                         
                         // 計算目標數量（當前明細中應該出貨的數量，以負數表示）
-                        int targetQuantity = hasCurrent ? -currentInventory[key].CurrentQuantity : 0;
+                        decimal targetQuantity = hasCurrent ? -currentInventory[key].CurrentQuantity : 0m;
                         
                         // 計算已處理的庫存數量（之前所有交易的淨值，已經是負數）
-                        int processedQuantity = hasProcessed ? processedInventory[key].NetProcessedQuantity : 0;
+                        decimal processedQuantity = hasProcessed ? processedInventory[key].NetProcessedQuantity : 0m;
                         
                         // 計算需要調整的數量
-                        int adjustmentNeeded = targetQuantity - processedQuantity;
+                        decimal adjustmentNeeded = targetQuantity - processedQuantity;
                         
                         if (adjustmentNeeded != 0)
                         {
